requirements.txt
streamlit
pandas
numpy
openpyxl
fpdf2
